requirements.txt
alembic==1.0.0
astroid==2.0.2
atomicwrites==1.1.5
attrs==18.1.0
click==6.7
dominate==2.3.1
flake8==3.5.0
Flask==1.0.2
Flask-Admin==1.5.1
Flask-Bootstrap4==4.0.2
Flask-Login==0.4.1
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
pluggy==0.7.1
py==1.5.4
pycodestyle==2.3.1
pyflakes==1.6.0
pylint==2.1.1
pytest==3.7.1
python-dateutil==2.7.3
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.10
typed-ast==1.1.0
typing==3.6.4
visitor==0.1.3
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
